fix: extract shared rendering into cmd/render.go (phase 2)
Move JSON types, display helpers, and rendering functions from up.go
into a dedicated render.go. Split printServiceLine boolean flag into
printServiceLineDetailed and printServiceLineCompact. Extract shared
serviceURLSuffix and portStatusSuffix helpers.

up.go: 520 → 328 lines. render.go: 213 lines.
